pyproject.toml
[tool.poetry]
name = "bookingsystemapi"
version = "0.1.0"
description = ""
authors = ["olszewskik <kamil.web@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.2"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
SQLAlchemy = "^1.4.35"
alembic = "^1.7.7"
PyMySQL = "^1.0.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
